Refactor doctor command
Separate the types of warnings (only 1 for now)
